feat: add comprehensive configuration options and streaming support for Google GenAI models, including `ModelInfo` and an integer `index` in `ModelResponseChunk`.

+++ packages/genkit/lib/src/ai/generate.dart
@@ -74,7 +74,7 @@ class GenerateResponseChunk<O> {
   });
 
   // Delegate properties to _chunk
-  double? get index => _chunk.index;
+  int? get index => _chunk.index;
   Role? get role => _chunk.role;
   List<Part> get content => _chunk.content;
   Map<String, dynamic>? get custom => _chunk.custom;
